Use real alert schema in retrieve_alerts tests
While the mock schema was simple, it didn't really check all the
corners of our code. In particular, it didn't test the behavior of the
SCHEMA_DEFS cache which is internal to fastavro. That cache gets used
when fastavro reads a complex record which uses named types, like
lsst.diaSource, by reference. We need to be extra-careful in how we
handle modifications to that cache while iterating over records, and
modifications are plausible because we modify it when we instantiate a
lsst.alert.packet.Schema wrapper around the
fastavro.reader.writer_schema in retrieve_alerts - that instantiation
causes the SCHEMA_DEFS cache to be cleared.

This commit just adds a test case that reveals the above problem, it
doesn't fix it.