Remove all old schemas and put version number into schema namespace
The way to hold multiple versions of an Avro schema at once is through
namespaces since fastavro uses a global map, SCHEMA_DEFS, to keep
track of types it has already seen and parsed. Previous versions of
lsst.alert.packet reused names and namespaces, which made it very
difficult to interoperate with fastavro.

This is the first commit in a small series. It deletes all the old
schemas and renames v3.0's schemas, and changes their internal
references.

We could rename v3.0 to v1.0.0, but that's more intrusive without
clear benefit. We could remove all of the directory structure (since
version information is right in the filename, so they won't collide),
but that would be more intrusive too.

Tests don't pass with this commit.