Replace pkg_resources with importlib.resources
The API here is possibly dangerous in that resources can
refer to ZIP files which might lead to the file path being
returned to a temporary file that will be deleted.

In all our common scenarios this is not an issue but a
real fix would be to allow a URI to be returned that can
be resolved later.